2_FastAPI_Service/requirements.txt

fastapi
uvicorn[standard]
pillow
requests
python-multipart

